Normalize Whiting thumbnails
The image normalization process modifies the image histogram so that the range of grey tones goes from black to white.  It is a lossless process so no information is lost.  The end result is that the images are more similar in tone range, i.e., none look too dark or to faded.  Images with deep shadows in pin holes may still look a bit faded compared to the rest but in general the images look like they were photographed under the same lighting conditions.  Normalization makes it easier to compare the tonal values of different grounds.